refactor: move sortable fields in table
